refactor: remove problems with lint
