+++ src/styles/style.scss
@@ -1,3 +1,2 @@
 @import 'base/base';
-@import 'abstracts/variables';
 @import 'components/keyboard';
